requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
bootstrap4==0.1.0
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
Django==2.2.4
django-bootstrap4==0.0.8
django-tastypie==0.14.2
feedparser==5.2.1
get==2019.4.13
idna==2.8
importlib-metadata==0.19
lxml==4.4.0
mock==3.0.5
model-mommy==1.6.0
more-itertools==7.2.0
nested-lookup==0.2.17
packaging==19.1
pluggy==0.12.0
post==2019.4.13
public==2019.4.13
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
wcwidth==0.1.7
xmljson==0.2.0
zipp==0.5.2
